+++ src/API.PeopleAdministrative.Application/Requests/UpdatePersonRequest.cs
@@ -6,14 +6,14 @@ namespace API.PeopleAdministrative.Application.Requests;
 
 public sealed class UpdatePersonRequest : BaseRequestWithValidation
 {
-    public UpdatePersonRequest(int cpf, string nome, string endereco)
+    public UpdatePersonRequest(long cpf, string nome, string endereco)
     {
         Cpf = cpf;
         Nome = nome;
         Endereco = endereco;
     }
 
-    public int Cpf { get; private set; }
+    public long Cpf { get; private set; }
     public string Nome { get; private set; }
     public string Endereco { get; private set; }
