+++ src/API.PeopleAdministrative.Domain/Entities/Person.cs
@@ -4,12 +4,12 @@ namespace API.PeopleAdministrative.Domain.Entities;
 
 public sealed class Person : IEntity
 {
-    public int Cpf { get; private set; }
+    public long Cpf { get; private set; }
     public string Nome { get; private set; }
     public string Endereco { get; private set; }
     public bool Ativo { get; private set; }
 
-    public Person SetCpf(int cpf)
+    public Person SetCpf(long cpf)
     {
         Cpf = cpf;
         return this;
