+++ src/API.PeopleAdministrative.Domain/Interfaces/IPeopleRepository.cs
@@ -9,5 +9,5 @@ public interface IPeopleRepository : IRepository
 {
     Task CreateAsync(Person people, CancellationToken cancellationToken = default);
     Task UpdateAsync(Person people, CancellationToken cancellationToken = default);
-    Task<Person> SearchAsync(int cpf, CancellationToken cancellationToken = default);
+    Task<Person> SearchAsync(long cpf, CancellationToken cancellationToken = default);
 }
